Add files via upload
This is the Python script I created to clean up the wine review data to fit what I wanted to do with it. The wine reviews I used are publicly available at https://www.kaggle.com/zynicide/wine-reviews , where the user "zackthoutt" scraped the data from www.winemag.com. This script extracts the first 1000 reviews, for a more manageable amount of data for the app. It cleans up the wine reviews into descriptor words mainly, while also classifying the wines as red, white, or sparkling/other, based on a lot of kinds of wine that I looked up and manually added in. It can be useful for others using the 130k reviews looking to extract the data into a format that is easier to work with, that can classify based on wine color, that also makes it easy to convert into JSON objects. I cleaned the data outside the app so it would save run time on the app.